+++ CHANGELOG.md
@@ -7,7 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-Nothing.
+### Added
+
+- The header title can now have an URL to point to.
+
+### TODO
+
+- class for metadata row div
+- fix author expanding too much
+- quote component
+- note component
 
 ## [0.0.2] - 2024-10-26