+++ CHANGELOG.md
@@ -10,10 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - The header title can now have an URL to point to.
+- The post metadata div now have an exposed class.
 
 ### TODO
 
-- class for metadata row div
 - fix author expanding too much
 - quote component
 - note component