+++ CHANGELOG.md
@@ -5,9 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.0] - 2024-10-27
 
-Nothing.
+### Added
+
+- The header title can now have an URL to point to.
+- The post metadata div now have an exposed class.
+- Added Quote post component.
+- Added framed text block component, supporting user defined variants.
+
+### Fixed
+
+- The aside cannot expand as much as it wants anymore.
+
+## [0.0.2] - 2024-10-26
+
+### Fixed
+
+- fixed module being empty on remote due to `dist` being untracked.
 
 ## [0.0.1] - 2024-10-26